Move test in core package
